re-write the [Run] section using inno-script
So we can control exactly when it runs. Now we only install
vbox or msys if they are missing from the system.

And thus updating vbox doesn't fall afoul of the 'cannot find kernel
driver' issue on upgrade, and msys doesn't fail because the explorer
addin is loaded.

much much YAY